Transition from long to long long

+++ main.c
@@ -19,10 +19,10 @@ int main(void) {
     // BUBBLE SORT RANDOM VALUES
     FILE* input_1_1 = fopen("data/data1.txt", "r"); // input random values from the file
     FILE* output_1_1 = fopen("data/output1.1.txt", "w");  // output file for bubble sort random input
-    long* a = (long*) malloc(n * sizeof(long));
+    long long* a = (long long*) malloc(n * sizeof(long long));
     for(int i = 0; i < n; i++)  // scan data from input file
     {
-        fscanf(input_1_1, "%ld", &a[i]);
+        fscanf(input_1_1, "%lld", &a[i]);
     }
     fprintf(res, "\nBubble sort random values\n");
     bubble_sort(a, n, count);
@@ -31,7 +31,7 @@ int main(void) {
     count[1] = 0; //
     for(int i = 0; i < n; i++) // print data to output file
     {
-        fprintf(output_1_1, "%ld\n", a[i]);
+        fprintf(output_1_1, "%lld\n", a[i]);
     }
     fclose(input_1_1);
     fclose(output_1_1);
@@ -40,7 +40,7 @@ int main(void) {
     FILE* output_1_2 = fopen("data/output1.2.txt", "w");  // output file for qsort sort random input
     for(int i = 0; i < n; i++)  // scan data from input file
     {
-        fscanf(input_1_2, "%ld", &a[i]);
+        fscanf(input_1_2, "%lld", &a[i]);
     }
     fprintf(res, "\n\nQuick sort random values\n");
     qsort_alg(a, 0, n, count);
@@ -49,7 +49,7 @@ int main(void) {
     count[1] = 0; //
     for(int i = 0; i < n; i++)
     {
-        fprintf(output_1_2, "%ld\n", a[i]);
+        fprintf(output_1_2, "%lld\n", a[i]);
     }
     fclose(input_1_2);
     fclose(output_1_2);
@@ -58,7 +58,7 @@ int main(void) {
     FILE* output_2_1 = fopen("data/output2.1.txt", "w"); // output for sorted file
     for(int i = 0; i < n; i++)  // scan data from input file
     {
-        fscanf(input_2_1, "%ld", &a[i]);
+        fscanf(input_2_1, "%lld", &a[i]);
     }
     fprintf(res, "\n\nBubble sort sorted n to 1 values\n");
     bubble_sort(a, n, count);
@@ -67,7 +67,7 @@ int main(void) {
     count[1] = 0; //
     for(int i = 0; i < n; i++)
     {
-        fprintf(output_2_1, "%ld\n", a[i]);
+        fprintf(output_2_1, "%lld\n", a[i]);
     }
     fclose(input_2_1);
     fclose(output_2_1);
@@ -76,7 +76,7 @@ int main(void) {
     FILE* output_2_2 = fopen("data/output2.2.txt", "w");
     for(int i = 0; i < n; i++)  // scan data from input file
     {
-        fscanf(input_2_2, "%ld", &a[i]);
+        fscanf(input_2_2, "%lld", &a[i]);
     }
     fprintf(res, "\n\nQuick sort sorted n to 1 values\n");
     qsort_alg(a, 0, n, count);
@@ -85,7 +85,7 @@ int main(void) {
     count[1] = 0; //
     for(int i = 0; i < n; i++)
     {
-        fprintf(output_2_2, "%ld\n", a[i]);
+        fprintf(output_2_2, "%lld\n", a[i]);
     }
     fclose(input_2_2);
     fclose(output_2_2);
@@ -94,7 +94,7 @@ int main(void) {
     FILE* output_3_1 = fopen("data/output3.1.txt", "w");
     for(int i = 0; i < n; i++)  // scan data from input file
     {
-        fscanf(input_3_1, "%ld", &a[i]);
+        fscanf(input_3_1, "%lld", &a[i]);
     }
     fprintf(res, "\n\nBubble sort sorted 1 to n values\n");
     bubble_sort(a, n, count);
@@ -103,7 +103,7 @@ int main(void) {
     count[1] = 0; //
     for(int i = 0; i < n; i++)
     {
-        fprintf(output_3_1, "%ld\n", a[i]);
+        fprintf(output_3_1, "%lld\n", a[i]);
     }
     fclose(input_3_1);
     fclose(output_3_1);
@@ -112,7 +112,7 @@ int main(void) {
     FILE* output_3_2 = fopen("data/output3.2.txt", "w");
     for(int i = 0; i < n; i++)  // scan data from input file
     {
-        fscanf(input_3_2, "%ld", &a[i]);
+        fscanf(input_3_2, "%lld", &a[i]);
     }
     fprintf(res, "\n\nQuick sort sorted 1 to n values\n");
     qsort_alg(a, 0, n, count);
@@ -121,7 +121,7 @@ int main(void) {
     count[1] = 0; //
     for(int i = 0; i < n; i++)
     {
-        fprintf(output_3_2, "%ld\n", a[i]);
+        fprintf(output_3_2, "%lld\n", a[i]);
     }
     fclose(input_3_2);
     fclose(output_3_2);
